backend/requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
networkx==3.0
obsws-python==1.4.1
pathspec==0.10.3
platformdirs==2.6.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-vlc==3.0.18121
pywin32==305
requests==2.28.2
six==1.16.0
tinydb==4.7.1
tomli==2.0.1
ujson==5.7.0
urllib3==1.26.14
waitress==2.1.2
websocket-client==1.4.2
Werkzeug==2.2.2
